+++ files/paths.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io/ioutil"
-	"os"
 	"path/filepath"
 )
 
@@ -30,19 +29,6 @@ func buildKeyDir(gitdir string) string {
 
 func buildKeyFileName(path string) string {
 	return filepath.Join(path, DefaultKeyFile)
-}
-
-func checkFileMode(name string, fileinfo os.FileInfo, expected os.FileMode) error {
-	mode := fileinfo.Mode().Perm()
-	if mode&^expected != 0 {
-		return fmt.Errorf("excessive rights on %s", name)
-	}
-
-	if mode&expected != expected {
-		return fmt.Errorf("insufficient rights on %s", name)
-	}
-
-	return nil
 }
 
 // ExchangeDir returns Key Exchange dir inside the git repo